pyproject.toml
[project]
name = "flickr8k-image-captioning"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "albumentations>=2.0.8",
    "fastapi>=0.135.3",
    "grad-cam>=1.5.5",
    "gradio>=6.12.0",
    "hf-transfer>=0.1.9",
    "huggingface-hub>=1.10.2",
    "kagglehub>=1.0.0",
    "matplotlib>=3.10.8",
    "nltk>=3.9.4",
    "numpy>=2.4.4",
    "optuna>=4.8.0",
    "pandas>=3.0.2",
    "pillow>=12.2.0",
    "pycocoevalcap>=1.2",
    "pydantic>=2.13.0",
    "seaborn>=0.13.2",
    "torch==2.5.1",
    "torchvision==0.20.1",
    "tqdm>=4.67.3",
    "uvicorn>=0.44.0",
    "wandb>=0.26.0",
    "wordcloud>=1.9.6",
]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
    "jupyter>=1.1.1",
    "markdown>=3.10.2",
    "plotly>=6.7.0",
    "pytest>=9.0.3",
    "python-pptx>=1.0.2",
    "reportlab>=4.4.10",
    "ruff>=0.15.10",
    "torchinfo>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src", "api"]

[tool.uv.sources]
torch = { index = "pytorch-cu121" }
torchvision = { index = "pytorch-cu121" }

[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true
